requirements.txt
pandas
python-docx
pdf2image
Pillow
pytesseract
pyarrow
openpyxl
openai
